requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.7.0
